requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pyyaml==6.0.1
feedparser==6.0.11
python-dateutil>=2.8.2
tqdm>=4.65.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
lxml>=4.9.3
html5lib>=1.1
fake-useragent>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.1
seaborn>=0.12.2
aiohttp>=3.8.5
langchain>=0.0.300
biopython>=1.81
validators>=0.20.0 